+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.9.0](https://github.com/Workvia/w_transport/compare/2.8.1...2.9.0)
+_TBD_
+
+- **Improvement:** All request classes now have a `bool isDone` getter that can
+  be read to determine whether or not a request is complete (i.e. succeeded,
+  failed, or canceled).
+
 ## [2.8.0](https://github.com/Workvia/w_transport/compare/2.7.1...2.8.0)
 _July 21, 2016_